fix: refactored EventDrawer
